backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
google-genai
pandas
openpyxl
python-dotenv
httpx